Move texas holdem type to own file
